Module for stealer e2e.
