+++ tests/src/traffic/steal.rs
@@ -1,0 +1,85 @@
+#[cfg(test)]
+mod steal {
+    use std::{net::UdpSocket, time::Duration};
+
+    use futures::Future;
+    use futures_util::stream::TryStreamExt;
+    use k8s_openapi::api::core::v1::Pod;
+    use kube::{api::LogParams, Api, Client};
+    use rstest::*;
+
+    #[cfg(target_os = "linux")]
+    use crate::utils::{get_service_url, send_requests, Agent, Application};
+    use crate::utils::{
+        kube_client, run_exec, service, udp_logger_service, KubeService, CONTAINER_NAME,
+    };
+
+    #[cfg(target_os = "linux")]
+    #[rstest]
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    #[timeout(Duration::from_secs(240))]
+    async fn test_steal_http_traffic(
+        #[future] service: KubeService,
+        #[future] kube_client: Client,
+        #[values(
+            Application::PythonFlaskHTTP,
+            Application::PythonFastApiHTTP,
+            Application::NodeHTTP
+        )]
+        application: Application,
+        #[values(Agent::Ephemeral, Agent::Job)] agent: Agent,
+    ) {
+        let service = service.await;
+        let kube_client = kube_client.await;
+        let url = get_service_url(kube_client.clone(), &service).await;
+        let mut flags = vec!["--steal"];
+        agent.flag().map(|flag| flags.extend(flag));
+        let mut process = application
+            .run(&service.target, Some(&service.namespace), Some(flags), None)
+            .await;
+
+        process.wait_for_line(Duration::from_secs(40), "daemon subscribed");
+        send_requests(&url, true).await;
+        tokio::time::timeout(Duration::from_secs(40), process.child.wait())
+            .await
+            .unwrap()
+            .unwrap();
+
+        application.assert(&process);
+    }
+
+    #[cfg(target_os = "linux")]
+    #[rstest]
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    #[timeout(Duration::from_secs(240))]
+    async fn test_steal_http_filter_matches_for_client_a(
+        #[future] service: KubeService,
+        #[future] kube_client: Client,
+        #[values(
+            Application::PythonFlaskHTTP,
+            Application::PythonFastApiHTTP,
+            Application::NodeHTTP
+        )]
+        application: Application,
+        #[values(Agent::Ephemeral, Agent::Job)] agent: Agent,
+    ) {
+        let service = service.await;
+        let kube_client = kube_client.await;
+        let url = get_service_url(kube_client.clone(), &service).await;
+        let mut flags = vec!["--steal"];
+        agent.flag().map(|flag| flags.extend(flag));
+        let mut client_a = application
+            // TODO(alex) [high] 2022-12-26: Insert env var for stealer filter.
+            .run(&service.target, Some(&service.namespace), Some(flags), None)
+            .await;
+
+        client_a.wait_for_line(Duration::from_secs(40), "daemon subscribed");
+        send_requests(&url, true).await;
+        tokio::time::timeout(Duration::from_secs(40), client_a.child.wait())
+            .await
+            .unwrap()
+            .unwrap();
+
+        application.assert(&client_a);
+    }
+}
